+++ com.ibm.streamsx.inet/impl/java/src/com/ibm/streamsx/inet/http/Oauth2Signer.java
@@ -1,0 +1,12 @@
+package com.ibm.streamsx.inet.http;
+
+public class Oauth2Signer {
+
+	private String scope = null;
+	private String grant_type = "client_credentials";
+	private String client_id = null;
+	private String client_secret = null;
+	private String access_token = null;
+	private String authentication_server_url = null;
+	
+}